feat: extract slideshow and load landscape photos for home
